refactor: separate Firebase and Local strategies, call based on config
