+++ api/server/services/Files/Local/index.js
@@ -1,0 +1,7 @@
+const images = require('./images');
+const save = require('./save');
+
+module.exports = {
+  ...save,
+  ...images,
+};
