+++ api/server/services/Files/images/index.js
@@ -1,4 +1,3 @@
-const convert = require('./convert');
 const encode = require('./encode');
 const parse = require('./parse');
 const resize = require('./resize');
@@ -6,7 +5,6 @@ const validate = require('./validate');
 const uploadAvatar = require('./avatar/uploadAvatar');
 
 module.exports = {
-  ...convert,
   ...encode,
   ...parse,
   ...resize,
