--- api/server/services/Files/index.js
@@ -1,9 +1,0 @@
-const localStrategy = require('./localStrategy');
-const process = require('./process');
-const save = require('./save');
-
-module.exports = {
-  ...save,
-  ...process,
-  localStrategy,
-};
